App Configuration
This app configuration class, sets up the core functionalities of the application. It utilizes the Get package for state management and navigation. The class defines the app's theme and routing configuration using the GetMaterialApp.router widget. It also handles the initialization and disposal of router and loading state instances. The app follows a dark theme mode and disables the debug banner for a polished user experience