CardSetAdapter.
The CardSetAdapter class is a generated code that serves as a type adapter for the CardSet model. It is part of the 'card_set.dart' file. The adapter allows reading and writing instances of the CardSet class to and from binary format. It implements the TypeAdapter interface and includes methods for reading and writing CardSet objects. The adapter is specifically designed for a CardSet object and defines the type identifier as 4.